+++ src/helpers/Sanitize.ts
@@ -1,6 +1,6 @@
-const illegalRe = /[/?<>\\:*|"]/g;
+const illegalRe = /[/?<>\\:*|"!.,;{}[\]()_+=~`@#$%^&]/g;
 // eslint-disable-next-line no-control-regex
-const controlRe = /[\x00-\x1F\x80-\x9F]/g;
+const controlRe = /[\x00-\x1f\x80-\x9f]/g;
 const reservedRe = /^\.+$/;
 const windowsReservedRe = /^(con|prn|aux|nul|com[0-9]|lpt[0-9])(\..*)?$/i;
 const windowsTrailingRe = /[. ]+$/;
@@ -9,6 +9,7 @@ function sanitize(input: string, replacement: string) {
   if (typeof input !== 'string') {
     throw new Error('Input must be string');
   }
+
   const sanitized = input
     .replace(illegalRe, replacement)
     .replace(controlRe, replacement)